requirements.txt
Django>=5.1.1,<5.2
djangorestframework>=3.15.2,<3.16
django-cors-headers>=4.0.0,<4.1
django-extensions>=3.2.1,<3.3
django-modeltranslation>=0.18.10,<0.19
psycopg2-binary>=2.9.9,<3.0
gunicorn>=22,<23
django-phonenumber-field[phonenumberslite]>=7.1.0,<7.2
django-silk>=5.2.0,<5.3
drf-yasg>=1.21.8,<1.22
pillow>=11.1.0, <11.2
